fix(android): make bridgeless compatible (#33)
As we are migrating our react native app to new arch we need to make our modules compatible with it. This module is quite outdated in terms of react-native API usage, however, the goal of this PR is to just make it work for now.

The main fixes are:

We can't use the catalystInstance (bridge) in bridgeless, so we had to replace all usages of it. It was only really used to directly invoke a JS callback (which didn't even seem to work reliably)
We removed the code path in shouldOverrideUrlLoading that would need to work on the catalystInstance. In our product code we don't use this functionality on android. As its code is quite wicked i just removed it and left a todo to properly replace it (then we can align android and iOS in product code as well which is neat)
We shouldn't do getNativeModule(UIManagerModule), i replaced that with the proper call to get the UI manager + fixed up the react tag retrieval